+++ test/routes_rebuild_test.dart
@@ -159,35 +159,6 @@ void main() {
     expect(routeMap1UnknownRoutes.isEmpty, isTrue);
     expect(routeMap2UnknownRoutes.isEmpty, isTrue);
   });
-
-  testWidgets("Doesn't change navigator key when route map changes",
-      (tester) async {
-    final delegate = RoutemasterDelegate(routesBuilder: (context) {
-      StateProvider.of(context).state;
-
-      return RouteMap(routes: {
-        '/': (_) => MaterialPage<void>(child: PageOne()),
-      });
-    });
-    final state = AppState();
-
-    await tester.pumpWidget(
-      StateProvider(
-        state: state,
-        child: MaterialApp.router(
-          routeInformationParser: RoutemasterParser(),
-          routerDelegate: delegate,
-        ),
-      ),
-    );
-
-    final navigatorKey1 = tester.widget(find.byType(Navigator)).key;
-    state.someValue = 'state change';
-    await tester.pump();
-    final navigatorKey2 = tester.widget(find.byType(Navigator)).key;
-
-    expect(navigatorKey1, navigatorKey2);
-  });
 }
 
 class AppState extends ChangeNotifier {