+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.9.0
+
+* Fixed: issue with rebuilding routes when `RoutemasterDelegate` is recreated 
+
+# 0.8.1
+
+* Fixed: widgets rebuilding when `RoutemasterDelegate` is recreated
+
 # 0.8.0
 
 * Breaking change: `Guard` properties have been renamed: `validate` is now `canNavigate` and `onValidationFailed` is now `onNavigationFailed`